Adds a 'Match after scan' option to the manual schedule configuration options for scanning.
