+++ client/components/modals/libraries/EditModal.vue
@@ -125,6 +125,7 @@ export default {
           skipMatchingMediaWithAsin: false,
           skipMatchingMediaWithIsbn: false,
           autoScanCronExpression: null,
+          matchAfterScan: false,
           hideSingleBookSeries: false,
           onlyShowLaterBooksInContinueSeries: false,
           metadataPrecedence: ['folderStructure', 'audioMetatags', 'nfoFile', 'txtFiles', 'opfFile', 'absMetadata'],
