+++ client/strings/en-us.json
@@ -448,6 +448,7 @@
   "LabelLookForNewEpisodesAfterDate": "Look for new episodes after this date",
   "LabelLowestPriority": "Lowest Priority",
   "LabelMatchConfidence": "Confidence",
+  "LabelMatchAfterScan": "Run 'Match Books' after scan",
   "LabelMatchExistingUsersBy": "Match existing users by",
   "LabelMatchExistingUsersByDescription": "Used for connecting existing users. Once connected, users will be matched by a unique id from your SSO provider",
   "LabelMaxEpisodesToDownload": "Max # of episodes to download. Use 0 for unlimited.",
