+++ server/models/Library.js
@@ -68,6 +68,7 @@ class Library extends Model {
         coverAspectRatio: 1, // Square
         disableWatcher: false,
         autoScanCronExpression: null,
+        matchAfterScan: false,
         skipMatchingMediaWithAsin: false,
         skipMatchingMediaWithIsbn: false,
         audiobooksOnly: false,
